chore: move zone navigation features
