+++ MacsyZones/DonationReminder.swift
@@ -178,25 +178,28 @@ class DonationReminder {
         
         if countI % interval == 0 {
             // Check if zone navigation is happening and suppress donation reminder
+            // This is because its a "spammy" action and this would be extremely annoying
+            // if we didnt skip it
             if isZoneNavigating {
                 return
             }
-                if #available(macOS 12.0, *) {
-                    if isQuickSnapping {
-                        self.panel.orderFront(nil)
+
+            if #available(macOS 12.0, *) {
+                if isQuickSnapping {
+                    self.panel.orderFront(nil)
                     // Only bring QuickSnapper to front if it's actually open
                     if quickSnapper.isOpen {
                         quickSnapper.panel.makeKeyAndOrderFront(nil)
                     } else {
                         debugLog("DonationReminder.count() - QuickSnapper is not open, not bringing it to front")
                     }
-                        self.panel.makeKeyAndOrderFront(nil)
-                    }
-                } else {
                     self.panel.makeKeyAndOrderFront(nil)
                 }
-                
-                self.panel.center()
+            } else {
+                self.panel.makeKeyAndOrderFront(nil)
+            }
+            
+            self.panel.center()
             
             countI = 0
         }
