+++ MacsyZones/Macsy.swift
@@ -25,7 +25,6 @@ var isFitting = false
 var isEditing = false
 var isQuickSnapping = false
 var isSnapResizing = false
-var isZoneNavigating = false
 
 var isMovingAWindow = false
 
@@ -210,10 +209,10 @@ func getHoveredSectionWindow() -> SectionWindow? {
                 }
             } else {
                 sortedSectionWindows = userLayouts.currentLayout.layoutWindow.sectionWindows.sorted {
-                    let frame1 = $0.window.frame
-                    let frame2 = $1.window.frame
-                    return (frame1.width * frame1.height) < (frame2.width * frame2.height)
-                }
+                let frame1 = $0.window.frame
+                let frame2 = $1.window.frame
+                return (frame1.width * frame1.height) < (frame2.width * frame2.height)
+            }
             }
             
             for sectionWindow in sortedSectionWindows {
@@ -430,6 +429,29 @@ func isElementResizable(element: AXUIElement) -> Bool {
     return resizable.boolValue
 }
 
+func snapWindowToZone(sectionNumber: Int, element: AXUIElement, windowId: UInt32) {
+    let section = userLayouts.currentLayout.layoutWindow.sectionWindows.first(where: { $0.sectionConfig.number! == sectionNumber })
+    if let section, let sectionWindow = section.window {
+        guard let (screenNumber, workspaceNumber) = SpaceLayoutPreferences.getCurrentScreenAndSpace() else { return }
+        guard let focusedScreen = NSScreen.screens.first(where: { $0.frame.contains(NSEvent.mouseLocation) }) else { return }
+        
+        if !PlacedWindows.isPlaced(windowId: windowId) { OriginalWindowProperties.update(windowID: windowId) }
+        
+        // Use immediate positioning for zone navigation to prevent flickering
+        moveWindowToMatch(element: element,
+                          targetWindow: sectionWindow,
+                          targetScreen: focusedScreen,
+                          sectionConfig: section.sectionConfig)
+        
+        PlacedWindows.place(windowId: windowId,
+                            screenNumber: screenNumber,
+                            workspaceNumber: workspaceNumber,
+                            layoutName: userLayouts.currentLayoutName,
+                            sectionNumber: sectionNumber,
+                            element: element)
+    }
+}
+
 func resizeAndMoveWindow(element: AXUIElement, newPosition: CGPoint, newSize: CGSize, retries: Int = 10, retryParent: Bool = false) {
     if retryParent && !isElementResizable(element: element) {
         debugLog("Window is not resizable! Trying parent window...")
@@ -462,8 +484,8 @@ func resizeAndMoveWindow(element: AXUIElement, newPosition: CGPoint, newSize: CG
     
     // Approach to reduce flickering:
     // 1. Immediate attempt (atomic set position + size)
-    // 2. If after a short delay the window drifted (position OR size changed noticeably), perform one corrective pass with light staggered retries
-    // We consider a drift if either axis differs more than 2px or size differs more than 2px (tolerates rapid sequential moves without over-correcting)
+    // 2. If after a short delay the window drifted perform one corrective pass with light staggered retries
+    // This is for cases where HTML/CSS built apps (Discord, Teams, Slack, etc) mess with the window
     var initialWindowId: UInt32? = getWindowID(from: element)
     var originalObserved: (CGSize?, CGPoint?) = (nil, nil)
     if let id = initialWindowId {
@@ -686,9 +708,9 @@ func onMouseUp(event: NSEvent) {
             
             justDidMouseUp = true
         }
-        
-        isFitting = false
-        userLayouts.currentLayout.layoutWindow.hide()
+            
+            isFitting = false
+            userLayouts.currentLayout.layoutWindow.hide()
     } else {
         isFitting = false
         toLeaveElement = nil
@@ -740,178 +762,6 @@ func cycleWindowsInZone(forward: Bool) {
     activateElementsWindow(element: targetWindow.element)
     
     debugLog("Cycled \(forward ? "forward" : "backward") to window \(targetWindow.windowId)")
-}
-
-// MARK: - Zone Navigation Functions
-func snapWindowToZone(sectionNumber: Int, element: AXUIElement, windowId: UInt32) {
-    let section = userLayouts.currentLayout.layoutWindow.sectionWindows.first(where: { $0.sectionConfig.number! == sectionNumber })
-    if let section, let sectionWindow = section.window {
-        guard let (screenNumber, workspaceNumber) = SpaceLayoutPreferences.getCurrentScreenAndSpace() else { return }
-        guard let focusedScreen = NSScreen.screens.first(where: { $0.frame.contains(NSEvent.mouseLocation) }) else { return }
-        
-        if !PlacedWindows.isPlaced(windowId: windowId) { OriginalWindowProperties.update(windowID: windowId) }
-        
-        // Use immediate positioning for zone navigation to prevent flickering
-    moveWindowToMatch(element: element, 
-               targetWindow: sectionWindow, 
-               targetScreen: focusedScreen, 
-               sectionConfig: section.sectionConfig)
-        
-        PlacedWindows.place(windowId: windowId,
-                            screenNumber: screenNumber,
-                            workspaceNumber: workspaceNumber,
-                            layoutName: userLayouts.currentLayoutName,
-                            sectionNumber: sectionNumber,
-                            element: element)
-    }
-}
-
-enum ZoneDirection {
-    case left, right, up, down
-}
-
-func moveWindowToAdjacentZone(direction: ZoneDirection) {
-    debugLog("moveWindowToAdjacentZone called with direction: \(direction)")
-    
-    // Set zone navigation flag to suppress donation reminders
-    isZoneNavigating = true
-    debugLog("moveWindowToAdjacentZone - Set isZoneNavigating=true")
-    
-    // Use defer to ensure flag is reset when function exits
-    defer {
-        isZoneNavigating = false
-        debugLog("moveWindowToAdjacentZone - Reset isZoneNavigating=false (defer)")
-    }
-    
-    guard let focusedElement = getFocusedWindowAXUIElement(),
-          let focusedWindowId = getWindowID(from: focusedElement) else {
-        debugLog("moveWindowToAdjacentZone - No focused window found for zone navigation")
-        return
-    }
-    
-    debugLog("moveWindowToAdjacentZone - Found focused window ID: \(focusedWindowId)")
-    
-    // Get current zone if window is already placed
-    let currentSectionNumber = PlacedWindows.isPlaced(windowId: focusedWindowId) ? 
-        PlacedWindows.windows[focusedWindowId] : nil
-    
-    // Find target zone based on direction
-    guard let targetSectionNumber = findAdjacentZone(
-        from: currentSectionNumber,
-        direction: direction,
-        windowElement: focusedElement
-    ) else {
-        debugLog("moveWindowToAdjacentZone - No adjacent zone found in \(direction) direction")
-        return
-    }
-    
-    debugLog("moveWindowToAdjacentZone - Moving window \(focusedWindowId) from zone \(currentSectionNumber?.description ?? "none") to zone \(targetSectionNumber)")
-    snapWindowToZone(sectionNumber: targetSectionNumber, element: focusedElement, windowId: focusedWindowId)
-}
-
-func findAdjacentZone(from currentSection: Int?, direction: ZoneDirection, windowElement: AXUIElement) -> Int? {
-    let sectionWindows = userLayouts.currentLayout.layoutWindow.sectionWindows
-    guard !sectionWindows.isEmpty else { return nil }
-    
-    // Get current window rect for reference
-    let currentRect: NSRect
-    if let currentSection = currentSection,
-       let currentSectionWindow = sectionWindows.first(where: { $0.number == currentSection }) {
-        currentRect = currentSectionWindow.sectionConfig.getRect()
-    } else {
-        // Use actual window position if not in a zone
-        var position = CGPoint.zero
-        var size = CGSize.zero
-        var positionRef: CFTypeRef?
-        var sizeRef: CFTypeRef?
-        
-        if AXUIElementCopyAttributeValue(windowElement, kAXPositionAttribute as CFString, &positionRef) == .success,
-           let positionValue = positionRef {
-            AXValueGetValue(positionValue as! AXValue, .cgPoint, &position)
-        }
-        
-        if AXUIElementCopyAttributeValue(windowElement, kAXSizeAttribute as CFString, &sizeRef) == .success,
-           let sizeValue = sizeRef {
-            AXValueGetValue(sizeValue as! AXValue, .cgSize, &size)
-        }
-        
-        currentRect = NSRect(origin: position, size: size)
-    }
-    
-    // Use edge-based logic for all other cases (including windows already in zones)
-    var bestZone: SectionWindow?
-    var bestDistance: CGFloat = CGFloat.greatestFiniteMagnitude
-    
-    let currentCenter = CGPoint(x: currentRect.midX, y: currentRect.midY)
-    
-    for sectionWindow in sectionWindows {
-        // Skip current zone
-        if sectionWindow.number == currentSection { continue }
-        
-        let zoneRect = sectionWindow.sectionConfig.getRect()
-        let zoneCenter = CGPoint(x: zoneRect.midX, y: zoneRect.midY)
-        
-        // Check if zone is in the correct direction
-        let isInDirection: Bool
-        let distance: CGFloat
-        
-        switch direction {
-        case .left:
-            // Zone must be to the left: zone's right edge should be to the left of window's left edge
-            isInDirection = (zoneRect.maxX <= currentRect.minX + 50) // Allow 50px overlap for better UX
-            if isInDirection {
-                // Distance from window's left edge to zone's right edge + vertical alignment penalty
-                let horizontalDistance = max(0, currentRect.minX - zoneRect.maxX)
-                let verticalDistance = abs(zoneCenter.y - currentCenter.y)
-                distance = horizontalDistance + (verticalDistance * 0.5) // Weight vertical alignment less
-            } else {
-                distance = CGFloat.greatestFiniteMagnitude
-            }
-            
-        case .right:
-            // Zone must be to the right: zone's left edge should be to the right of window's right edge
-            isInDirection = (zoneRect.minX >= currentRect.maxX - 50) // Allow 50px overlap for better UX
-            if isInDirection {
-                // Distance from window's right edge to zone's left edge + vertical alignment penalty
-                let horizontalDistance = max(0, zoneRect.minX - currentRect.maxX)
-                let verticalDistance = abs(zoneCenter.y - currentCenter.y)
-                distance = horizontalDistance + (verticalDistance * 0.5) // Weight vertical alignment less
-            } else {
-                distance = CGFloat.greatestFiniteMagnitude
-            }
-            
-        case .up:
-            // Zone must be above: zone's bottom edge should be above window's top edge
-            isInDirection = (zoneRect.maxY <= currentRect.minY + 50) // Allow 50px overlap for better UX
-            if isInDirection {
-                // Distance from window's top edge to zone's bottom edge + horizontal alignment penalty
-                let verticalDistance = max(0, currentRect.minY - zoneRect.maxY)
-                let horizontalDistance = abs(zoneCenter.x - currentCenter.x)
-                distance = verticalDistance + (horizontalDistance * 0.5) // Weight horizontal alignment less
-            } else {
-                distance = CGFloat.greatestFiniteMagnitude
-            }
-            
-        case .down:
-            // Zone must be below: zone's top edge should be below window's bottom edge
-            isInDirection = (zoneRect.minY >= currentRect.maxY - 50) // Allow 50px overlap for better UX
-            if isInDirection {
-                // Distance from window's bottom edge to zone's top edge + horizontal alignment penalty
-                let verticalDistance = max(0, zoneRect.minY - currentRect.maxY)
-                let horizontalDistance = abs(zoneCenter.x - currentCenter.x)
-                distance = verticalDistance + (horizontalDistance * 0.5) // Weight horizontal alignment less
-            } else {
-                distance = CGFloat.greatestFiniteMagnitude
-            }
-        }
-        
-        if isInDirection && distance < bestDistance {
-            bestDistance = distance
-            bestZone = sectionWindow
-        }
-    }
-    
-    return bestZone?.number
 }
 
 func getWindowsInSameZone(as windowId: UInt32) -> [(windowId: UInt32, element: AXUIElement)] {
@@ -976,6 +826,6 @@ func presentingShortcut(_ shortcut: String) -> String {
         .replacingOccurrences(of: "Control", with: "⌃")
         .replacingOccurrences(of: "Option", with: "⌥")
         .replacingOccurrences(of: "Shift", with: "⇧")
-    
+
     return formattedShortcut
 }
