+++ MacsyZones/Onboarding.swift
@@ -378,6 +378,7 @@ struct OnboardingView: View {
 }
 
 @available(macOS 12.0, *)
+@available(macOS 12.0, *)
 struct OnboardingPageView: View {
     let page: OnboardingPage
     @State private var player: AVPlayer?
@@ -518,6 +519,7 @@ struct OnboardingPageView: View {
 }
 
 @available(macOS 12.0, *)
+@available(macOS 12.0, *)
 struct OnboardingWindowView: View {
     @ObservedObject var state = onboardingState
     @State private var showOnboarding = false
@@ -539,6 +541,7 @@ struct OnboardingWindowView: View {
 
 private var onboardingWindow: NSWindow?
 
+@available(macOS 12.0, *)
 @available(macOS 12.0, *)
 func showOnboarding() {
     let window = NSWindow()
@@ -568,9 +571,8 @@ func showOnboarding() {
     window.makeKeyAndOrderFront(nil)
 }
 
+@available(macOS 12.0, *)
 #Preview {
-    if #available(macOS 12.0, *) {
-        OnboardingView(window: nil)
-    }
+    OnboardingView(window: nil)
 }
